requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
jinja2==3.1.4
python-multipart==0.0.22
